Fix obj parser Increase the size of the box

+++ ouzel/scene/ModelData.cpp
@@ -22,7 +22,7 @@ namespace ouzel
         }
 
         bool ModelData::init(Box3 newBoundingBox,
-                             const std::vector<uint16_t> indices,
+                             const std::vector<uint32_t> indices,
                              const std::vector<graphics::VertexPCT>& vertices,
                              const std::shared_ptr<graphics::Material>& newMaterial)
         {
@@ -35,7 +35,7 @@ namespace ouzel
             vertexBuffer->init(graphics::Buffer::Usage::VERTEX, vertices.data(), static_cast<uint32_t>(getVectorSize(vertices)));
 
             meshBuffer = std::make_shared<graphics::MeshBuffer>();
-            meshBuffer->init(sizeof(uint16_t), indexBuffer, graphics::VertexPCT::ATTRIBUTES, vertexBuffer);
+            meshBuffer->init(sizeof(uint32_t), indexBuffer, graphics::VertexPCT::ATTRIBUTES, vertexBuffer);
 
             material = newMaterial;
